bpf: Add xattr support to bpffs
Introduces bpf_inode_info as a container for the vfs_inode.
This container allows for additional information about the
inode to be stored in bpffs, specifically security xattrs.

Additionaly adds a kmem_cache bpf_inode_cache to store
these containers.

This work to is required to enable SELinux relabelling of
parts of a bpffs - support was previously added to label
pre-defined paths in SELinux policy using genfscon

Signed-off-by: Dave Tucker <dave@dtucker.co.uk>